requirements.txt
requests==2.25.1
pytest==6.2.2
responses==0.12.1
requests-mock==1.8.0
Flask==1.1.2
pact-python==1.3.0